refactor: module integer - Changes to be committed: -	new file:   src/integer.mjs   - feat: asInteger to convert a value into an integer.   - feat: isInteger to test a value is an integer.   - feat: type Int for numbers, which are integers.   - feat: type Integer for number and bigint integers. -	modified:   src/temporal.mjs   - refactor: replaced asInteger and isInteger with imports.   - feat: export type Int   - feat: export type Integer
